+++ GlyssenCharacters/CharacterVerseData.cs
@@ -116,7 +116,7 @@ namespace GlyssenCharacters
 
 		public static bool IsCharacterExtraBiblical(string characterId)
 		{
-			// See comment in IsCharacterStandard regarding checking the length first as a the speed optimization
+			// See comment in IsCharacterStandard regarding checking the length first as a speed optimization
 			switch (characterId?.Length)
 			{
 				case 6: return characterId.StartsWith(kBookOrChapterPrefix, StringComparison.Ordinal);
@@ -127,7 +127,7 @@ namespace GlyssenCharacters
 
 		public static bool IsInterruption(string characterId)
 		{
-			// See comment in IsCharacterStandard regarding checking the length first as a the speed optimization
+			// See comment in IsCharacterStandard regarding checking the length first as a speed optimization
 			return characterId?.Length == 16 && characterId.StartsWith(kInterruptionPrefix, StringComparison.Ordinal);
 		}